Forgot length of type
git-svn-id: svn+ssh://svn.phoenixbox.net/svn/v6502/trunk@659 96dfbd92-d197-e211-9ca9-00110a534b34

+++ as6502/error.c
@@ -42,7 +42,7 @@ __attribute((noreturn)) void as6502_fatal(const char *reason) {
 	exit(EXIT_FAILURE);
 }
 
-void _setProblemLocation(unsigned long loc, unsigned len) {
+void _setProblemLocation(unsigned long loc, unsigned long len) {
 	lastProblematicLine = currentLineNum;
 
 	startOfProblem = loc;
